Revert "uint int comparison fixes in rendermodel.cpp"
This reverts commit 13ff07973a8ab1e15ae02ad1e1b40377c8581f73.

+++ src/engine/render/rendermodel.cpp
@@ -682,7 +682,7 @@ void resetmodelbatches()
 void addbatchedmodel(model *m, batchedmodel &bm, int idx)
 {
     modelbatch *b = nullptr;
-    if(static_cast<int>(batches.size()) > m->batch)
+    if(batches.size() > m->batch)
     {
         b = &batches[m->batch];
         if(b->m == m && (b->flags & Model_Mapmodel) == (bm.flags & Model_Mapmodel))
@@ -1094,7 +1094,7 @@ void startmodelquery(occludequery *query)
 {
     modelquery = query;
     modelquerybatches = batches.size();
-    modelquerymodels = static_cast<int>(batchedmodels.size());
+    modelquerymodels = batchedmodels.size();
     modelqueryattached = modelattached.size();
 }
